Changement des parametres JpaRepository

+++ src/main/java/com/univtime/informatique/repositories/PromoEstComposeeRepository.java
@@ -3,6 +3,6 @@ package com.univtime.informatique.repositories;
 import com.univtime.informatique.entities.PromoEstComposeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromoEstComposeeRepository extends JpaRepository<PromoEstComposeeEntity, Long> {
+public interface PromoEstComposeeRepository extends JpaRepository<PromoEstComposeeEntity, Integer> {
 
 }
